DBC22-4684: refactoring, styling, fixed cam orientation carousel
