+++ src/frontend/src/Components/cameras/directions/CameraOrientations.scss
@@ -9,7 +9,6 @@
   @media (min-width: 1440px) {
     min-width: 400px;
   }
-  
 
   .header {
     display: flex;
@@ -75,23 +74,31 @@
     }
 
     @media (min-width: 1440px) {
-      column-count: 2;
-      column-gap: 12px;
-      padding: 8px 0;
+      display: flex;
+      flex-flow: wrap;
+      width: 432px;
+      column-count: unset;
     }
 
     .orientation {
       position: relative;
-      height: 160px;
-      width: 100%;
+      height: 132px;
       border-radius: 8px;
       overflow: hidden;
       margin-bottom: 12px;
       cursor: pointer;
+      display: flex;
+      align-content: center;
+      justify-content: center;
+      background: black;
 
       @media (min-width: 768px) {
         width: 210px;
         height: 160px;  // Crop off 20px timestamp from the bottom
+
+        &:nth-child(odd) {
+          margin-right: 12px;
+        }
       }
 
       @media (min-width: 992px) {
@@ -109,14 +116,9 @@
       }
 
       img {
-        height: calc(100% + 20px);
         width: 100%;
-        object-position: top;
-
-        @media (min-width: 768px) {
-          width: 210px;
-          height: 180px;
-        }
+        height: auto;
+        clip-path: inset(0 0 5px 0); /* Crops timestamp */
       }
 
       .label {
@@ -162,13 +164,13 @@
         }
 
         &.next {
-          right: 0;
+          right: 4px;
           animation: show 600ms 100ms cubic-bezier(0.38, 0.97, 0.56, 0.76) forwards;
         }
 
         &.prev {
-          left: 0;
-          animation: show 600ms 100ms cubic-bezier(0.38, 0.97, 0.56, 0.76) forwards;      
+          left: 4px;
+          animation: show 600ms 100ms cubic-bezier(0.38, 0.97, 0.56, 0.76) forwards;
         }
       }
     }
